+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from "./AssetConfiguration";
+export * from "./AssetBundleConfiguration";
 export * from "./GameConfiguration";
 export * from "./OperationPluginInfo";